fix(github-actions): automerge all non-major github actions with stable version
Previously, patch and minor version updates for Github actions would never match the "all non-major dependencies with stable version" grouping because Github action versions are prefixed with a "v". This change adds a new group that is specific to github actions that expects tags to be prefixed with a "v" (i.e. "v1.0.0").